LayoutTests/fast/security-labels/inner-test-expected.txt

This tests that setting textContent and innerText propagates labels properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS tag.isOn(target.textContent) is false
PASS tag.isOn(target.innerText) is false
PASS tag.isOn(target.textContent) is true
PASS tag.isOn(target.innerText) is true
PASS tag.isOn(target.innerText) is true
PASS tag.isOn(target.innerHTML) is true
PASS successfullyParsed is true

TEST COMPLETE
foo
